refactor(ts-migration): move policies to typescript (#1670)
## Summary
- add the policies migration phase manifest
- mechanically move `bin/lib/policies.js` into `src/lib/policies.ts`
- keep a compatibility wrapper in `bin/lib/policies.js`
- rename `test/policies.test.js` to `.ts`

## Testing
- npm run build:cli
- npm run typecheck:cli
- npm run lint
- npx vitest run test/policies.test.ts
- npm test
